requirements.txt
google-api-python-client==2.120.0
youtube-transcript-api>=1.2.4
anthropic>=0.83.0
matplotlib==3.8.4
Pillow==10.3.0
requests>=2.31.0
python-dotenv==1.0.1
numpy==1.26.4
